refactor: phase-6 structural splits (ad_detector, schema, AdReviewModal)
Three big-file decompositions from the audit. All HIGH-risk per
the plan. Strategy was pure code-motion only -- git mv for the
central renames so blame survives, no behavior change, no signature
or log-line edits, no SQL DDL byte changes.

PR 6.1: src/ad_detector.py -> src/ad_detector/ package.
- __init__.py 1403 LOC: AdDetector class + explicit named re-exports
  covering all 28 public names tests and production import (audited
  via `grep "from ad_detector import"` repo-wide, including the
  underscored ones tests reach into directly: _NON_BRAND_WORDS,
  _extract_ad_keywords, _find_keyword_region, _unpack_region,
  _text_has_ad_content, _find_json_array_candidates).
- boundaries.py 857 LOC: refine/extend/snap/merge/validate/dedupe
  helpers + transition phrases.
- prompts.py 393 LOC: USER_PROMPT_TEMPLATE, create_windows,
  format_window_prompt, get_static_system_prompt,
  parse_ads_from_response.
- Central file dropped 2575 -> 1403 LOC (-45%). Net +78 LOC from
  added module docstrings and duplicated imports.
- Git records: R src/ad_detector.py -> src/ad_detector/__init__.py.

PR 6.2: src/database/schema.py -> src/database/schema/ package.
EXTREMELY HIGH risk per CLAUDE.md "never lose data on migrations".
Pure code motion only. SHA-256 of SCHEMA_SQL and
MIGRATION_INDEXES_SQL identical before/after the split
(13aac307..., a73d3d14...). No MIGRATIONS registry exists in this
codebase to reorder; the migration sequence is encoded in a single
770-line _run_schema_migrations method that was not touched. Only
the two SQL DDL string constants moved out to tables.py (306 LOC);
re-imported from __init__.py for back-compat. SchemaMixin class
stays in __init__.py.
- Migration-focused tests (test_migration_sponsor_fk,
  test_only_expose_processed_migration, test_settings_migration,
  test_database): 92 passed.

PR 6.3: frontend/src/components/AdReviewModal.tsx decomposition.
3 of 6 audit-planned extractions:
- frontend/src/utils/adReviewHelpers.ts: state-free helpers
  (parseTimeInput, formatTime, getThemeWaveformColors,
  loadPlayWhileDragging, savePlayWhileDragging, PLAY_WHILE_DRAG_KEY).
- frontend/src/components/ad-editor/Pin.tsx: Pin subcomponent +
  PinProps interface. Self-contained, owns its own dragging state.
  MIN_AD_DURATION=1.0 duplicated locally to avoid a circular import;
  comment in Pin.tsx flags the two-site invariant.
- frontend/src/components/ad-editor/usePeaks.ts: peaks fetch/cancel
  effect + the three peaks state vars. setPeaks(null) on resetView
  dropped because the hook already clears+re-fetches on resetTick.
- AdReviewModal.tsx: 1654 -> 1448 LOC (-12.5%). Default export and
  named exports unchanged.

Scope-limited per the audit's "if extraction causes subtle
reactivity changes, revert" guard:
- useWavesurfer NOT extracted -- wavesurfer mount deliberately omits
  adStart/adEnd from deps (eslint-disable); wrapping would either
  reintroduce deps (changes reactivity) or require ref-passing that
  complicates the surface.
- AdReviewToolbar + Cursor + usePinDrag NOT extracted -- prop counts
  per component would balloon to 8-15 for marginal LOC savings.

Verification: 1177 pytest passed, frontend tsc clean.
